unify database driver handling
