+++ internal/dbdrivers/sqlite/driver.go
@@ -5,6 +5,11 @@ package sqlite
 import (
 	"context"
 	"database/sql/driver"
+	"fmt"
+	"net/url"
+	"os"
+	"os/exec"
+	"strings"
 
 	"github.com/arran4/goa4web/internal/dbdrivers"
 	"github.com/mattn/go-sqlite3"
@@ -35,6 +40,48 @@ func (c connector) Driver() driver.Driver { return &sqlite3.SQLiteDriver{} }
 
 func (Driver) OpenConnector(dsn string) (driver.Connector, error) {
 	return connector{dsn: dsn}, nil
+}
+
+func pathFromConn(conn string) string {
+	path := conn
+	if strings.HasPrefix(path, "file:") {
+		if u, err := url.Parse(path); err == nil {
+			path = u.Path
+		}
+	}
+	return path
+}
+
+// Backup dumps the SQLite database to file using the sqlite3 command.
+func (Driver) Backup(dsn, file string) error {
+	path := pathFromConn(dsn)
+	cmd := exec.Command("sqlite3", path, ".dump")
+	outFile, err := os.Create(file)
+	if err != nil {
+		return fmt.Errorf("create file: %w", err)
+	}
+	defer outFile.Close()
+	cmd.Stdout = outFile
+	if err := cmd.Run(); err != nil {
+		return fmt.Errorf("backup: %w", err)
+	}
+	return nil
+}
+
+// Restore loads the SQLite database from file using the sqlite3 command.
+func (Driver) Restore(dsn, file string) error {
+	path := pathFromConn(dsn)
+	inFile, err := os.Open(file)
+	if err != nil {
+		return fmt.Errorf("open file: %w", err)
+	}
+	defer inFile.Close()
+	cmd := exec.Command("sqlite3", path)
+	cmd.Stdin = inFile
+	if err := cmd.Run(); err != nil {
+		return fmt.Errorf("sqlite restore: %w", err)
+	}
+	return nil
 }
 
 // Register registers the SQLite driver.
